requirements.txt
streamlit>=1.33.0
pydantic>=2.0
supabase-py>=2.0
SQLAlchemy[asyncpg]>=2.0
python-dotenv>=1.0.0
streamlit-authenticator>=0.3.0
ruff
passlib[bcrypt]
pytest
pytest-asyncio
